pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "openwebui-mcp-server"
version = "0.1.0"
description = "External MCP Server for OpenWebUI Agent Handoff"
requires-python = ">=3.8"
dependencies = [
    "mcp>=1.4.1",
    "httpx>=0.28.1",
    "starlette>=0.46.1", 
    "uvicorn>=0.34.0",
    "python-dotenv>=1.0.0"
]